Update README.md
Rearchitected the content to separate the concepts and implementation tasks. The quick start guide now includes all the pertinent tasks for installation and configuration. I created a table with links for users to easily access all the topics in the quick start guide and simplified the content for consistency and ease of use. 
This is a draft version. Therefore there is plenty of room for improvement.